Fix: cast of unsigned int to int

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -21,7 +21,7 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 	{
 		x = va_arg(myargs, int);
 		separator != NULL ? printf("%d%s", x,
-		i == n - 1 ? "\n" : separator) : printf("%d", x);
+		i == (int) n - 1 ? "\n" : separator) : printf("%d", x);
 		i++;
 	}
 	if (separator == NULL)
